Moved layout.js to the components folder, changed the font of the title and fixed the title of the pages and the title of the page that appears in the search engine.
