+++ src/components/layout.module.css
@@ -1,8 +1,12 @@
 
+.body {
+  background-color: red;
+}
+
 .container {
   margin: auto;
   max-width: 500px;
-  font-family: Arial;
+  font-family: Helvetica Neue;
 }
 
 .heading {
