+++ src/components/seo.js
@@ -1,19 +1,20 @@
 import * as React from 'react'
 import { graphql, useStaticQuery } from 'gatsby'
 
-const Seo = ({ title }) => {
+const Seo = ({ title, pageTitle }) => {
   const data = useStaticQuery(graphql`
     query {
       site {
         siteMetadata {
           title
+          pageTitle
         }
       }
     }
   `)
 
   return (
-    <title>{title} | {data.site.siteMetadata.title}</title>
+    <title>{data.site.siteMetadata.pageTitle}</title>
   )
 }
