+++ src/pages/about.js
@@ -1,6 +1,6 @@
 // Step 1: Import React
 import * as React from 'react'
-import Layout from '../pages/layout'
+import Layout from '../components/layout'
 import Seo from '../components/seo'
 
 // Step 2: Define your component
@@ -12,10 +12,8 @@ const AboutPage = () => {
   )
 }
 
-// export const Head = () => <title>About Me</title>
-
 export const Head = () => (
-  <Seo title="About Page" description="Troy's about page. Here, you'll learn how cool I am."/>
+  <Seo title="Student | Troy Lin - About" pageTitle="About" description="Troy's about page. Here, you'll learn how cool I am."/>
 )
 
 // Step 3: Export your component
