+++ src/pages/blog/index.js
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import Layout from '../../pages/layout'
+import Layout from '../../components/layout'
 import { Link, graphql } from 'gatsby'
 import Seo from '../../components/seo'
 
@@ -37,6 +37,6 @@ export const query = graphql`
   }
 `
 
-export const Head = () => <Seo title="My Blog Posts" />
+export const Head = () => <Seo title="Troy Lin | Personal Website | Blog" pageTitle="Blog" description="Hello I'm Troy Lin, and here's a collection of updates on important things :)"/>
 
 export default BlogPage
